refactor: OpenSearchIndex를 인프라 패키지로 이동
