+++ src/main/java/com/bookspot/batch/job/BookOpenSearchSyncJobConfig.java
@@ -1,6 +1,6 @@
 package com.bookspot.batch.job;
 
-import com.bookspot.batch.global.config.OpenSearchIndex;
+import com.bookspot.batch.infra.opensearch.OpenSearchIndex;
 import com.bookspot.batch.job.extractor.CommonStringJobParamExtractor;
 import com.bookspot.batch.job.listener.alert.AlertJobListener;
 import com.bookspot.batch.step.service.opensearch.OpenSearchRepository;
